+++ cmd/template/cluster/flag.go
@@ -22,9 +22,11 @@ const (
 	flagAWSControlPlaneSubnet = "control-plane-subnet"
 
 	// Cluster App only.
+	flagClusterAppCatalog        = "cluster-app-catalog"
 	flagClusterAppVersion        = "cluster-app-version"
 	flagClusterUserConfigMap     = "cluster-user-configmap"
 	flagClusterTopology          = "cluster-topology"
+	flagDefaultAppsAppCatalog    = "default-apps-app-catalog"
 	flagDefaultAppsAppVersion    = "default-apps-app-version"
 	flagDefaultAppsUserConfigMap = "default-apps-user-configmap"
 
@@ -76,11 +78,15 @@ type openStackFlag struct {
 }
 
 type clusterAppFlag struct {
-	ClusterUserConfigMap     string
-	ClusterAppVersion        string
-	ClusterTopology          bool
-	DefaultAppsUserConfigMap string
+	ClusterTopology bool
+
+	ClusterAppCatalog    string
+	ClusterAppVersion    string
+	ClusterUserConfigMap string
+
 	DefaultAppsAppVersion    string
+	DefaultAppsAppCatalog    string
+	DefaultAppsUserConfigMap string
 }
 
 type flag struct {
@@ -129,10 +135,12 @@ func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&f.OpenStack.NodeCIDR, flagOpenStackNodeCIDR, "", "CIDR used for the nodes.")
 
 	// OpenStack App only.
+	cmd.Flags().StringVar(&f.ClusterApp.ClusterAppCatalog, flagClusterAppCatalog, "giantswarm", "Cluster App version to be installed. (OpenStack App CR only).")
+	cmd.Flags().StringVar(&f.ClusterApp.ClusterAppVersion, flagClusterAppVersion, "0.3.0", "Cluster App version to be installed. (OpenStack App CR only).")
 	cmd.Flags().StringVar(&f.ClusterApp.ClusterUserConfigMap, flagClusterUserConfigMap, "", "Path to the user values configmap YAML file for Cluster App (OpenStack App CR only).")
-	cmd.Flags().StringVar(&f.ClusterApp.ClusterAppVersion, flagClusterAppVersion, "0.2.1", "Cluster App version to be installed. (OpenStack App CR only).")
-	cmd.Flags().StringVar(&f.ClusterApp.DefaultAppsUserConfigMap, flagDefaultAppsUserConfigMap, "", "Path to the user values configmap YAML file for Default Apps App (OpenStack App CR only).")
+	cmd.Flags().StringVar(&f.ClusterApp.DefaultAppsAppCatalog, flagDefaultAppsAppCatalog, "giantswarm", "Default Apps App version to be installed. (OpenStack App CR only).")
 	cmd.Flags().StringVar(&f.ClusterApp.DefaultAppsAppVersion, flagDefaultAppsAppVersion, "0.1.0", "Default Apps App version to be installed. (OpenStack App CR only).")
+	cmd.Flags().StringVar(&f.ClusterApp.DefaultAppsUserConfigMap, flagDefaultAppsUserConfigMap, "", "Path to the user values configmap YAML file for Default Apps App (OpenStack App CR only).")
 	cmd.Flags().BoolVar(&f.ClusterApp.ClusterTopology, flagClusterTopology, false, "Templated cluster as an App CR. (OpenStack App CR only).")
 
 	// TODO: Make these flags visible once we have a better method for displaying provider-specific flags.
@@ -149,9 +157,11 @@ func (f *flag) Init(cmd *cobra.Command) {
 	_ = cmd.Flags().MarkHidden(flagOpenStackNodeCIDR)
 
 	_ = cmd.Flags().MarkHidden(flagClusterTopology)
+	_ = cmd.Flags().MarkHidden(flagClusterAppCatalog)
 	_ = cmd.Flags().MarkHidden(flagClusterAppVersion)
-	_ = cmd.Flags().MarkHidden(flagDefaultAppsAppVersion)
 	_ = cmd.Flags().MarkHidden(flagClusterUserConfigMap)
+	_ = cmd.Flags().MarkHidden(flagDefaultAppsAppCatalog)
+	_ = cmd.Flags().MarkHidden(flagDefaultAppsAppVersion)
 	_ = cmd.Flags().MarkHidden(flagDefaultAppsUserConfigMap)
 
 	// Common.