+++ pkg/template/app/app.go
@@ -1,8 +1,11 @@
 package app
 
 import (
+	"fmt"
+
 	"github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
 	applicationv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
+	"github.com/giantswarm/k8smetadata/pkg/label"
 	"github.com/giantswarm/microerror"
 	"github.com/spf13/afero"
 	corev1 "k8s.io/api/core/v1"
@@ -24,6 +27,7 @@ type Config struct {
 	NamespaceConfigLabels      map[string]string
 	UserConfigConfigMapName    string
 	UserConfigSecretName       string
+	Organization               string
 	Version                    string
 }
 
@@ -41,25 +45,34 @@ type AppCROutput struct {
 
 func NewAppCR(config Config) ([]byte, error) {
 	userConfig := applicationv1alpha1.AppSpecUserConfig{}
+	appLabels := map[string]string{}
 
-	var namespace string
+	// Accomodating all the label cases here:
+	// 1. In-cluster Apps get unique label
+	// 2. Org-namespaced Apps get cluster label
+	// 3. Cluster-namespaced Apps with defaulting enabled gets nothing
+	var crNamespace string
 	if config.InCluster {
-		namespace = config.Namespace
+		crNamespace = config.Namespace
+		appLabels[label.AppOperatorVersion] = "0.0.0"
+	} else if config.Organization != "" {
+		crNamespace = fmt.Sprintf("org-%s", config.Organization)
+		appLabels[label.Cluster] = config.Cluster
 	} else {
-		namespace = config.Cluster
+		crNamespace = config.Cluster
 	}
 
 	if config.UserConfigConfigMapName != "" {
 		userConfig.ConfigMap = applicationv1alpha1.AppSpecUserConfigConfigMap{
 			Name:      config.UserConfigConfigMapName,
-			Namespace: namespace,
+			Namespace: crNamespace,
 		}
 	}
 
 	if config.UserConfigSecretName != "" {
 		userConfig.Secret = applicationv1alpha1.AppSpecUserConfigSecret{
 			Name:      config.UserConfigSecretName,
-			Namespace: namespace,
+			Namespace: crNamespace,
 		}
 	}
 
@@ -70,7 +83,8 @@ func NewAppCR(config Config) ([]byte, error) {
 		},
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      config.AppName,
-			Namespace: namespace,
+			Namespace: crNamespace,
+			Labels:    appLabels,
 		},
 		Spec: applicationv1alpha1.AppSpec{
 			Catalog:   config.Catalog,
@@ -88,21 +102,11 @@ func NewAppCR(config Config) ([]byte, error) {
 		},
 	}
 
-	if config.InCluster {
-		appCR.SetLabels(map[string]string{
-			"app-operator.giantswarm.io/version": "0.0.0",
-		})
-	}
-
 	if !config.DefaultingEnabled && !config.InCluster {
-		appCR.SetLabels(map[string]string{
-			"app-operator.giantswarm.io/version": "1.0.0",
-		})
-
 		appCR.Spec.Config = applicationv1alpha1.AppSpecConfig{
 			ConfigMap: applicationv1alpha1.AppSpecConfigConfigMap{
 				Name:      config.Cluster + "-cluster-values",
-				Namespace: config.Cluster,
+				Namespace: crNamespace,
 			},
 		}
 		appCR.Spec.KubeConfig = applicationv1alpha1.AppSpecKubeConfig{
@@ -112,7 +116,7 @@ func NewAppCR(config Config) ([]byte, error) {
 			InCluster: false,
 			Secret: applicationv1alpha1.AppSpecKubeConfigSecret{
 				Name:      config.Cluster + "-kubeconfig",
-				Namespace: config.Cluster,
+				Namespace: crNamespace,
 			},
 		}
 	}