+++ CHANGELOG.md
@@ -9,6 +9,18 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Added
 
+- Add support for templating App CRs in organization namespace.
+
+## [1.60.0] - 2022-01-27
+
+### Changed
+
+- Use `v1beta1` api version when templating ClusterAPI manifests on Azure.
+
+## [1.59.0] - 2022-01-26
+
+### Added
+
 - Add support to `template cluster --provider openstack` for templating clusters as App CRs.
 
 ## [1.58.2] - 2022-01-13
@@ -703,7 +715,9 @@ This release supports rendering for CRs:
 - `AppCatalog`
 - `App`
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v1.58.2...HEAD
+[Unreleased]: https://github.com/giantswarm/giantswarm/compare/v1.60.0...HEAD
+[1.60.0]: https://github.com/giantswarm/giantswarm/compare/v1.59.0...v1.60.0
+[1.59.0]: https://github.com/giantswarm/giantswarm/compare/v1.58.2...v1.59.0
 [1.58.2]: https://github.com/giantswarm/kubectl-gs/compare/v1.58.1...v1.58.2
 [1.58.1]: https://github.com/giantswarm/kubectl-gs/compare/v1.58.0...v1.58.1
 [1.58.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.57.0...v1.58.0